GPL
I couldn't figure out how to add another commit in my former issue (that is now merged, and maybe then too late?).. Sending as separate pull request. Seems ok, not to combine with the other (or combine if possible).

Note, I didn't say GPLv2 as I'm not sure that is what GPL means there. I guess this vagueness, is ok, people will at least then know and can look up (the file, I was to lazy too). If if is GPLv2 then, be default it can be upgraded to GPLv3 (unless it would say "only").

Only if it says GPLv3, there could be a problem (as not backwards compatible with GPLv2) but I doubt they do that, and anyway, it would strictly not be *your* problem.. You would at least be giving heads-up.