Actively check whether there's pending close notifications
Without this HTTP would try to push a request into a connection
that had already received a TLS close notify, which is an error
condition. Of course ideally HTTP would be robust to such error
conditions, but at least without a retry layer, it isn't. However,
just this patch reduces the incidence of HTTP errors to zero when
connecting to S3 (in combination with my PRs to HTTP.jl).