pk/rsa: Fix GC rooting issues and fill out API
MbedTLS is passing raw pointers to C without rooting ownership
all over the place. This fixes these issues in the pk/rsa code
and fills out the API a bit for round-trip testing, as well
as adding GC-safe wrappers over the internal mp integers for
downstream code to use. Of course, the unsafe GC pattern is
repeated elsewhere in this package, so this is just a first
PR that fixed the API surface that I happened to need.